+++ ORControllerClient/ORControllerClient/GridLayoutContainer.h
@@ -23,9 +23,14 @@
 @interface GridLayoutContainer : LayoutContainer
 
 @property (nonatomic, strong, readonly) NSMutableArray *cells;
-@property (nonatomic, readonly) int rows;
-@property (nonatomic, readonly) int cols;
+@property (nonatomic, readonly) NSUInteger rows;
+@property (nonatomic, readonly) NSUInteger cols;
 
-- (id)initWithLeft:(int)leftPos top:(int)topPos width:(int)widthDim height:(int)heightDim rows:(int)rowsNum cols:(int)colsNum;
+- (id)initWithLeft:(NSInteger)leftPos
+               top:(NSInteger)topPos
+             width:(NSUInteger)widthDim
+            height:(NSUInteger)heightDim
+              rows:(NSUInteger)rowsNum
+              cols:(NSUInteger)colsNum;
 
 @end
